ref(browser): Restructure browser-utils by domain
Reorganize browser-utils/src away from the catch-all metrics/ folder into
clear domains:

- instrumentation/ for event-source hooks (dom, history, location, xhr) plus
  the PerformanceObserver layer, renamed from metrics/instrument.ts to
  instrumentation/performanceObserver.ts to stop colliding with the folder
- web-vitals/ for web-vital tracking, spans, inp, lcp, reportEvents, and the
  browser helpers merged into a single utils.ts
- performance/ for generic performance-entry enrichment: entries, element/
  user/resource timing, and shared utils

browserMetrics.ts is split into web-vitals/tracking.ts and performance/
entries.ts, and utils.ts into web-vitals/reportEvents.ts and performance/
utils.ts. Public API (index.ts re-exports) is unchanged; only internal paths
and the test tree move.
