+++ packages/browser-utils/test/instrumentation/xhr.test.ts
@@ -1,6 +1,6 @@
 import type { HandlerDataXhr } from '@sentry/core';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { addXhrInstrumentationHandler, instrumentXHR } from '../../src/instrument/xhr';
+import { addXhrInstrumentationHandler, instrumentXHR } from '../../src/instrumentation/xhr';
 import { WINDOW } from '../../src/types';
 
 const win = WINDOW as typeof WINDOW & { XMLHttpRequest?: typeof XMLHttpRequest };
